+++ src/main/java/agents/plantOperator/PowerPlantOperator.java
@@ -7,7 +7,6 @@ import java.util.ArrayList;
 import java.util.List;
 import accounting.AnnualCostCalculator;
 import agents.trader.Trader;
-import agents.trader.TraderWithClients;
 import communications.message.AmountAtTime;
 import de.dlr.gitlab.fame.agent.Agent;
 import de.dlr.gitlab.fame.agent.input.DataProvider;
@@ -67,9 +66,9 @@ public abstract class PowerPlantOperator extends Agent {
 		ParameterData input = parameters.join(dataProvider);
 		annualCost = AnnualCostCalculator.build(input, "Refinancing");
 
-		call(this::executeDispatch).on(TraderWithClients.Products.DispatchAssignment)
-				.use(TraderWithClients.Products.DispatchAssignment);
-		call(this::digestPayment).on(TraderWithClients.Products.Payout).use(TraderWithClients.Products.Payout);
+		call(this::executeDispatch).on(PowerPlantScheduler.Products.DispatchAssignment)
+				.use(PowerPlantScheduler.Products.DispatchAssignment);
+		call(this::digestPayment).on(PowerPlantScheduler.Products.Payout).use(PowerPlantScheduler.Products.Payout);
 		call(this::reportCosts).on(Products.AnnualCostReport);
 	}