+++ src/main/java/agents/trader/renewable/AggregatorTrader.java
@@ -15,6 +15,7 @@ import agents.forecast.PowerForecastError;
 import agents.markets.DayAheadMarket;
 import agents.markets.DayAheadMarketTrader;
 import agents.plantOperator.PowerPlantOperator;
+import agents.plantOperator.PowerPlantScheduler;
 import agents.plantOperator.RenewablePlantOperator;
 import agents.plantOperator.RenewablePlantOperator.SetType;
 import agents.policy.SupportPolicy;
@@ -49,7 +50,7 @@ import de.dlr.gitlab.fame.time.TimeStamp;
 /** Aggregates supply capacity and administers support payments to plant operators
  * 
  * @author Johannes Kochems, Christoph Schimeczek, Felix Nitsch, Farzad Sarfarazi, Kristina Nienhaus */
-public abstract class AggregatorTrader extends TraderWithClients {
+public abstract class AggregatorTrader extends TraderWithClients implements PowerPlantScheduler {
 	@Input private static final Tree parameters = Make.newTree().addAs("ForecastError", PowerForecastError.parameters)
 			.buildTree();
 
@@ -111,10 +112,10 @@ public abstract class AggregatorTrader extends TraderWithClients {
 				.use(PowerPlantOperator.Products.MarginalCostForecast);
 		call(this::prepareBids).on(DayAheadMarketTrader.Products.Bids).use(PowerPlantOperator.Products.MarginalCost);
 		call(this::sendYieldPotentials).on(Products.YieldPotential).use(DayAheadMarket.Products.GateClosureInfo);
-		call(this::assignDispatch).on(TraderWithClients.Products.DispatchAssignment).use(DayAheadMarket.Products.Awards);
+		call(this::assignDispatch).on(PowerPlantScheduler.Products.DispatchAssignment).use(DayAheadMarket.Products.Awards);
 		call(this::requestSupportPayout).on(Products.SupportPayoutRequest);
 		call(this::digestSupportPayout).on(SupportPolicy.Products.SupportPayout).use(SupportPolicy.Products.SupportPayout);
-		call(this::payoutClients).on(TraderWithClients.Products.Payout);
+		call(this::payoutClients).on(PowerPlantScheduler.Products.Payout);
 	}
 
 	/** Extract information on {@link TechnologySet} and add it to the client data collection